+++ ChangeLog.md
@@ -2,9 +2,11 @@
 
 ## Unreleased changes
 
-- add support for different SAT solvers
-
 ## Released changes
+
+### 0.4.1
+
+- add support for different SAT solvers
 
 ### 0.4.0.3